gap_summary {HRAnalytics}	R Documentation

Get mean, median, and count where a column has values, split by a grouping factor

Description

Get mean, median, and count where a column has values, split by a grouping factor

Usage


gap_summary(df, group, column)

Arguments

df
Table to be analysed

group
Column to be grouped

column
Column to be aggregated

Value

tibble Summarising results

Examples


gap_summary(iris, Species, Sepal.Width)

[Package HRAnalytics version 0.0.0.9000 Index]