google_distance {HRAnalytics}	R Documentation

Get distance data between two points based on all the travel mode options. Works for many origin points.

Description

Get distance data between two points based on all the travel mode options. Works for many origin points.

Usage


google_distance(x, dest, arrival_time, key, ...)

Arguments

x
A vector of origins in address or postcode format

dest
A single destinationin address or postocde format

arrival_time
A POSIXct datetime that folks need to arrive by

key
A google distance API key

...
Additional options to pass to 'google_distance()'

Value

Data.frame containing (typically) 4 rows per input element

[Package HRAnalytics version 0.0.0.9000 Index]